% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_url.R
\name{build_url}
\alias{build_url}
\title{Construct a URL for the specified query.}
\usage{
build_url(base, query, start, end, step, timeout = NA)
}
\arguments{
\item{base}{A hostname and schema to base the generated path off of.}

\item{query}{A PromQL query.}

\item{start}{A RFC3339 timestamp string, numerical unix timestamp, or POSIXct
object.}

\item{end}{A RFC3339 timestamp string, numerical unix timestamp, or POSIXct
object.}

\item{step}{A query resolution step width.}

\item{timeout}{An optional query timeout value, defaulting to server-side
limit. Note this timeout is capped to the server-side value.}
}
\value{
A URL to execute the query.
}
\description{
Construct a URL for the specified query.
}
