% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_url.R
\name{cast_timestamp}
\alias{cast_timestamp}
\title{A helper function to map an input of various types to a timestamp string
suitable for use with Prometheus.}
\usage{
cast_timestamp(input)
}
\arguments{
\item{input}{A RFC3339 timestamp string, numerical unix timestamp, or POSIXct
object.}
}
\value{
A Prometheus-compatible timestamp that can be coerced to a string.
}
\description{
A helper function to map an input of various types to a timestamp string
suitable for use with Prometheus.
}
