% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_metrics_data.R
\name{mertic_formatting}
\alias{mertic_formatting}
\alias{format_metrics_instant_data}
\alias{format_metrics_range_data}
\alias{format_metadata}
\title{Various metric formatting function}
\usage{
format_metrics_instant_data(x)

format_metrics_range_data(x)

format_metadata(x)
}
\arguments{
\item{x}{A data frame generated by Prometheus query}
}
\value{
Formatted metrics object
}
\description{
Various metric formatting function
}
\section{Utility function for parsing metrics output}{

  Formats metrics data from the current query
}

\section{Utility function for parsing range metrics output}{

  Formats metrics data passed by range query.
}

\section{Utility function for parsing metadata output}{

  Formats data passed by metadata query.
}

