% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_to_xts.R
\name{metric_to_xts}
\alias{metric_to_xts}
\title{Convert Prometheus metric object to xts}
\usage{
metric_to_xts(metricsRaw)
}
\arguments{
\item{metricsRaw}{Raw instant or range query metrics object}
}
\value{
An xts object
}
\description{
The function is called via Prometheus query when argument other that
  \code{format = "raw"} is passed to the query call. Usually there is no
  need to call this function directly.
}
\details{
The function assumes that the passed object conforms to the raw
  Prometheus metric object
}
\examples{
\dontrun{
metrics_range <- prom$rangeQuery(query = "go_goroutines",
                                 start = as.numeric(as.POSIXct(Sys.time() - 600)),
                                 end = as.numeric(as.POSIXct(Sys.time())), step = "10s")
promR:::metric_to_xts(metrics_range)
# Equivalent to
#' metrics_range <- prom$rangeQuery(query = "go_goroutines",
                                 start = as.numeric(as.POSIXct(Sys.time() - 600)),
                                 end = as.numeric(as.POSIXct(Sys.time())), step = "10s",
                                 format = "xts")
}
}
