% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metrics_current}
\alias{metrics_current}
\title{Sample Promethous instant query results}
\format{
An object of class \code{data.frame} with 4 rows and 6 columns.
}
\usage{
metrics_current
}
\description{
The data set contains sample query result generated via
  \code{\link[promR]{Prometheus_query}} for a historical date. The data is provided
  by \url{demo.robustperception.io}.
}
\keyword{datasets}
