% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metrics_range}
\alias{metrics_range}
\title{Sample  Promethous range query results}
\format{An object of class \code{data.frame} with 120 rows and 6 columns.}
\usage{
metrics_range
}
\description{
The data set contains sample query result generated via
  \code{\link[promR]{Prometheus_rangeQuery}} for a historical time frame.
  The data is provided by \url{demo.robustperception.io}.
}
\keyword{datasets}
