% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promR.R
\docType{class}
\name{Prometheus-class}
\alias{Prometheus-class}
\alias{Prometheus}
\title{Prometheus class}
\description{
Reference Class to interface with a Prometheus Server
}
\section{Fields}{

\describe{
\item{\code{host}}{The hostname of the Prometheus server}

\item{\code{port}}{The port number of the Prometheus server}
}}

\section{Methods}{

\describe{
\item{\code{query(query, time = NULL, timeout = NULL)}}{Run an instant query}

\item{\code{rangeQuery(query, start, end, step, timeout = NULL)}}{Query data for a specific range}
}}

\examples{
prom <- Prometheus$new(host = "https://foo.bar", port = 9090)
}
