% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promR.R
\docType{class}
\name{Prometheus}
\alias{Prometheus}
\title{Prometheus}
\description{
Reference Class to interface with a Prometheus Server
}
\section{Fields}{

\describe{
\item{\code{host}}{The hostname of the Prometheus server}

\item{\code{port}}{The port number of the Prometheus server}
}}


\examples{
prom <- Prometheus$new(host = "https://foo.bar", port = 9090)
}
