% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promR.R
\name{Prometheus_metadataQuery}
\alias{Prometheus_metadataQuery}
\title{Prometheus Metadata Query}
\arguments{
\item{match_target}{The argument to match metrics against.}

\item{metric}{The metric to retrieve metadata for. Optional.}

\item{limit}{The number of results of targets to return.}
}
\description{
Run a metadata query against the Prometheus Server
}
\examples{
\dontrun{
prom <- Prometheus$new(host = "https://foo.bar", port = 9090)
metadata <- prom$metadataQuery(match_target = '{job=~"..*"}', metric = 'go_goroutines')
}
}
