% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promR.R
\name{Prometheus_rangeQuery}
\alias{Prometheus_rangeQuery}
\title{Promtheus Range Query}
\arguments{
\item{query}{The PromQL query}

\item{start}{Start timestamp, can be a rfc3339 or Unix timestamp.}

\item{end}{End timestamp, can be a rfc3339 or Unix timestamp.}

\item{step}{Query resolution step width (e.g. 15s)}

\item{timeout}{Evaluation timeout. Optional, defaults to timeout value of the Prometheus server.}
}
\description{
Run a range query against the Prometheus Server
}
\examples{
\dontrun{
prom <- Prometheus$new(host = "https://foo.bar", port = 9090)
rangeMetrics <-
 prom$rangeQuery(
   query = "up",
   start = "2018-10-25T20:10:51.781Z",
   end = "2018-10-26T20:10:51.781Z",
   step = "15s"
 )
}
}
