% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promR-package.R
\docType{package}
\name{promR-package}
\alias{promR}
\alias{promR-package}
\title{promR}
\description{
Package creating Prometheus class.

\if{html}{\figure{logo.png}{options: alt="promR logo" align="right"}}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/glenn-m/promR}
}

}
\author{
\strong{Maintainer}: Glenn McDonald \email{me@glennm.xyz}

Authors:
\itemize{
  \item Tyler Littlefield \email{tylurp1@gmail.com}
  \item Konrad Zdeb \email{konrad.zdeb@me.com} (0000-0002-4497-5681)
}

}
