% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{response_check}
\alias{response_check}
\title{Utility function for checking Prometheus Server response.}
\usage{
response_check(response)
}
\arguments{
\item{response}{httr response object}
}
\description{
Utility function for checking Prometheus Server response.
}
