% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-str.R
\name{string-utilities}
\alias{string-utilities}
\alias{rename_metadata_data_frame}
\alias{destring}
\title{String formatting functions}
\usage{
rename_metadata_data_frame(x)

destring(x)
}
\arguments{
\item{x}{A character vector}
}
\value{
Formatted string or destringed object.
}
\description{
String formatting functions
}
\section{Utility function cleaning column names}{

  As there is no clarity with respect to the returned columns the function:
  \itemize{
    \item Changes names of known columns like \code{target.instance, target.job}
    \item Removes special characters from all columns
  }
}

\section{Utility function used to destring value column}{

  The \code{value} columns should always contain numerical values, the
  function ensures that those are returned as of \code{numeric} type.
}

\keyword{internal}
