% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_timeout.R, R/utils.R
\name{parse_timeout}
\alias{parse_timeout}
\alias{utilities}
\title{Utility function for parsing timeout argument \code{parse_timeout}}
\usage{
parse_timeout(timeout_val)
}
\arguments{
\item{timeout_val}{A timeout value assigned by user.}
}
\description{
Utility function for parsing timeout argument \code{parse_timeout}

Basic arithmetic
}
