% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_metrics_data.R, R/parse_timeout.R,
%   R/utils.R
\name{format_metrics_instant_data}
\alias{format_metrics_instant_data}
\alias{format_metrics_range_data}
\alias{parse_timeout}
\alias{utilities}
\alias{response_check}
\title{Various utility functions}
\usage{
format_metrics_instant_data(x)

format_metrics_range_data(x)

parse_timeout(timeout_val)

response_check(response)
}
\arguments{
\item{x}{A data frame generated by Prometheus query}

\item{timeout_val}{A timeout value assigned by user.}

\item{response}{httr response object}
}
\description{
Utility functions, usually called within the package and not
  exported.
}
\section{Utility function for parsing metrics output}{

  Formats metrics data from the current query
}

\section{Utility function for parsing range metrics output}{

  Formats metrics data passed by range query.
}

\section{Utility function for parsing timeout argument \code{parse_timeout}}{

  Fixes potentially problematic timeout argument.
}

\section{Checking Prometheus Server response}{

  Returns more meanigful error code
}

