% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_metrics_data.R, R/parse_timeout.R,
%   R/utils.R
\name{format_metrics_data}
\alias{format_metrics_data}
\alias{parse_timeout}
\alias{utilities}
\title{Utility function for parsing timeout argument \code{parse_timeout}}
\usage{
format_metrics_data(x)

parse_timeout(timeout_val)
}
\arguments{
\item{x}{A data frame generated by Prometheus query}

\item{timeout_val}{A timeout value assigned by user.}
}
\description{
Utility function for parsing timeout argument \code{parse_timeout}

Utility function for parsing timeout argument \code{parse_timeout}

Basic arithmetic
}
