% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_metrics_data.R, R/parse_timeout.R,
%   R/utils.R
\name{format_metrics_instant_data}
\alias{format_metrics_instant_data}
\alias{format_metrics_range_data}
\alias{rename_metrics_data_frame}
\alias{destring}
\alias{parse_timeout}
\alias{utilities}
\alias{response_check}
\alias{metrics_check}
\title{Various utility functions}
\usage{
format_metrics_instant_data(x)

format_metrics_range_data(x)

rename_metrics_data_frame(x)

destring(x)

parse_timeout(timeout_val)

response_check(response)

metrics_check(metric)
}
\arguments{
\item{x}{A data frame generated by Prometheus query}

\item{timeout_val}{A timeout value assigned by user.}

\item{response}{httr response object}

\item{metrics}{Generated metrics object}
}
\description{
Utility functions, usually called within the package and not
  exported.
}
\section{Utility function for parsing metrics output}{

  Formats metrics data from the current query
}

\section{Utility function for parsing range metrics output}{

  Formats metrics data passed by range query.
}

\section{Utility function cleaning column names}{

  As there is no clarity with respect to the returned columns the function:
  \itemize{
    \item Chaneges names of known columns like \code{name}
    \item Removes special characters from all columns
  }
}

\section{Utility function used to destring value column}{

  The \code{value} columns should always contain numerical values, the
  function ensures that those are returned as of \code{numeric} type.
}

\section{Utility function for parsing timeout argument \code{parse_timeout}}{

  Fixes potentially problematic timeout argument.
}

\section{Checking Prometheus Server response}{

  Returns more meanigful error code
}

\section{Check Metrics Object}{

  Check if metrics objects contains data. If there is no data for time
  period referenced in the query this function will return an error.
}

\keyword{internal}
