% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_timeout.R, R/utils.R
\name{parse_timeout}
\alias{parse_timeout}
\alias{utilities}
\alias{response_check}
\alias{metrics_check}
\title{Various utility functions}
\usage{
parse_timeout(timeout_val)

response_check(response)

metrics_check(metric)
}
\arguments{
\item{timeout_val}{A timeout value assigned by user.}

\item{response}{httr response object}

\item{metrics}{Generated metrics object}
}
\description{
Utility functions, usually called within the package and not
  exported.
}
\section{Utility function for parsing timeout argument \code{parse_timeout}}{

  Fixes potentially problematic timeout argument.
}

\section{Checking Prometheus Server response}{

  Returns more meaningful error code
}

\section{Check Metrics Object}{

  Check if metrics objects contains data. If there is no data for time
  period referenced in the query this function will return an error.
}

\keyword{internal}
